+++ Package.swift
@@ -17,8 +17,8 @@ let package = Package(
     targets: [
         .binaryTarget(
             name: "AspirinShot",
-            url: "https://binaries.aspirinshot.com/aspirinshot/0.5/AspirinShot.zip",
-            checksum: "99c9141c14007bfc780363cfa3f4b059b22564a7d7bb391913748a49a781d268"
+            url: "https://binaries.aspirinshot.com/aspirinshot/0.6/AspirinShot.zip",
+            checksum: "714a969d2730365dd2be61530c4ff5ea61d0e979d29e498512cd3197f4ce18b5"
         )
     ]
 )